requirements.txt
astroid==2.3.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
future==0.18.2
geocoder==1.38.1
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.15.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.18.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
oauth2client==4.1.3
passlib==1.7.2
Pillow==7.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pypng==0.0.20
PyQRCode==1.2.1
pytz==2020.1
pyzbar==0.1.8
ratelim==0.1.6
requests==2.23.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.11.2


